refactor(util): add new package 'util' to improve project structure
- Store '*LottoGenerator.kt' files into 'util' package.
