+++ src/main/kotlin/lotto/util/LottoGenerator.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.util
 
 interface LottoGenerator {
     fun generateNumbers(): List<Int>
