+++ src/main/kotlin/lotto/util/RandomLottoGenerator.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.util
 
 import camp.nextstep.edu.missionutils.Randoms
